requirements.txt
wikipedia
langchain_openai
langchainhub
langchain_community 
langchain_openai
langchain_text_splitters 
langchainhub
wikipedia